refactor(channels): split feishu-channel.ts into modular components (Issue #694)
Split the 1059-line feishu-channel.ts into smaller, focused modules:

- passive-mode.ts (~60 lines): Manages passive mode state for group chats
- mention-detector.ts (~120 lines): Handles bot mention detection
- welcome-handler.ts (~130 lines): Manages welcome messages for P2P and group joins
- message-handler.ts (~470 lines): Handles incoming message events

The main feishu-channel.ts is now reduced to ~490 lines and serves as an
orchestrator that delegates to the modular components.

All 1572 tests pass after the refactoring.
