+++ src/channels/feishu-channel.ts
@@ -3,40 +3,41 @@
  *
  * Handles Feishu/Lark messaging platform integration via WebSocket.
  * Implements the IChannel interface for unified message handling.
+ *
+ * Issue #694: Refactored to use modular components
  */
 
 import * as lark from '@larksuiteoapi/node-sdk';
 import { Config } from '../config/index.js';
-import { DEDUPLICATION, REACTIONS, CHAT_HISTORY } from '../config/constants.js';
+import { REACTIONS } from '../config/constants.js';
 import { createLogger } from '../utils/logger.js';
-import { attachmentManager, downloadFile } from '../file-transfer/inbound/index.js';
+import { attachmentManager } from '../file-transfer/inbound/index.js';
 import { messageLogger } from '../feishu/message-logger.js';
 import { FeishuFileHandler } from '../platforms/feishu/feishu-file-handler.js';
 import { FeishuMessageSender } from '../platforms/feishu/feishu-message-sender.js';
 import { InteractionManager } from '../platforms/feishu/interaction-manager.js';
 import { createFeishuClient } from '../platforms/feishu/create-feishu-client.js';
 import type { WelcomeService } from '../platforms/feishu/welcome-service.js';
-import { getCommandRegistry } from '../nodes/commands/command-registry.js';
 import { resolvePendingInteraction } from '../mcp/feishu-context-mcp.js';
 import { TaskFlowOrchestrator } from '../feishu/task-flow-orchestrator.js';
 import { filteredMessageForwarder } from '../feishu/filtered-message-forwarder.js';
-import type { FilterReason } from '../config/types.js';
 import { TaskTracker } from '../utils/task-tracker.js';
-import { stripLeadingMentions } from '../utils/mention-parser.js';
 import { BaseChannel } from './base-channel.js';
 import type {
-  FeishuEventData,
-  FeishuMessageEvent,
-  FeishuCardActionEvent,
   FeishuCardActionEventData,
-  FeishuChatMemberAddedEventData,
-  FeishuP2PChatEnteredEventData,
+  FeishuCardActionEvent,
 } from '../types/platform.js';
 import type {
   ChannelConfig,
   OutgoingMessage,
   ChannelCapabilities,
 } from './types.js';
+
+// Import modular components (Issue #694)
+import { PassiveModeManager } from './feishu/passive-mode.js';
+import { MentionDetector } from './feishu/mention-detector.js';
+import { WelcomeHandler } from './feishu/welcome-handler.js';
+import { FeishuMessageHandler, type FeishuMessageHandlerContext } from './feishu/message-handler.js';
 
 const logger = createLogger('FeishuChannel');
 
@@ -71,32 +72,12 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
   private taskTracker: TaskTracker;
   private taskFlowOrchestrator?: TaskFlowOrchestrator;
   private interactionManager: InteractionManager;
-  private welcomeService?: WelcomeService;
 
-  private readonly MAX_MESSAGE_AGE = DEDUPLICATION.MAX_MESSAGE_AGE;
-
-  /**
-   * Passive mode state storage.
-   * Key: chatId, Value: true if passive mode is disabled (bot responds to all messages)
-   * Issue #511: Group chat passive mode control
-   */
-  private passiveModeDisabled: Map<string, boolean> = new Map();
-
-  /**
-   * Bot info for mention detection.
-   * Issue #600: Correctly identify bot mentions in group chats
-   * Issue #681: 群聊被动模式 @机器人检测不可靠问题
-   *
-   * Based on Feishu official documentation:
-   * - bot/v3/info returns bot.open_id and bot.app_id
-   * - When bot is mentioned, mentions[].id.open_id may be bot's open_id or app_id
-   *
-   * @see https://open.feishu.cn/document/uAjLw4CM/ukTMukTMukTM/reference/bot-v3/bot_info/get
-   */
-  private botInfo?: {
-    open_id: string;
-    app_id: string;
-  };
+  // Modular components (Issue #694)
+  private passiveModeManager: PassiveModeManager;
+  private mentionDetector: MentionDetector;
+  private welcomeHandler: WelcomeHandler;
+  private feishuMessageHandler: FeishuMessageHandler;
 
   constructor(config: FeishuChannelConfig = {}) {
     super(config, 'feishu', 'Feishu');
@@ -113,6 +94,7 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
           return { success: false };
         }
         try {
+          const { downloadFile } = await import('../file-transfer/inbound/index.js');
           const filePath = await downloadFile(this.client, fileKey, messageType, fileName, messageId);
           return { success: true, filePath };
         } catch (error) {
@@ -125,6 +107,13 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
     // Initialize InteractionManager
     this.interactionManager = new InteractionManager();
 
+    // Initialize modular components (Issue #694)
+    this.passiveModeManager = new PassiveModeManager();
+    this.mentionDetector = new MentionDetector();
+    this.welcomeHandler = new WelcomeHandler(this.appId, () => this.isRunning);
+    this.welcomeHandler.setRunningChecker(() => this.isRunning);
+    this.feishuMessageHandler = new FeishuMessageHandler();
+
     logger.info({ id: this.id }, 'FeishuChannel created');
   }
 
@@ -132,14 +121,18 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
     // Initialize message logger
     await messageLogger.init();
 
-    // Get bot info for mention detection (Issue #600, #681)
-    await this.fetchBotInfo();
+    // Get bot info for mention detection
+    this.getClient();
+    await this.mentionDetector.fetchBotInfo(this.client!);
 
     // Create event dispatcher
     this.eventDispatcher = new lark.EventDispatcher({}).register({
       'im.message.receive_v1': async (data: unknown) => {
         try {
-          await this.handleMessageReceive(data as FeishuEventData);
+          await this.feishuMessageHandler.handleMessageReceive(
+            data as any,
+            this.createMessageHandlerContext()
+          );
         } catch (error) {
           logger.error({ err: error }, 'Failed to handle message receive');
         }
@@ -152,18 +145,16 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
         }
       },
       'im.message.message_read_v1': async () => {},
-      // Issue #463: Handle P2P chat entered for welcome message
       'im.chat.access_event.bot_p2p_chat_entered_v1': async (data: unknown) => {
         try {
-          await this.handleP2PChatEntered(data as FeishuP2PChatEnteredEventData);
+          await this.welcomeHandler.handleP2PChatEntered(data as any);
         } catch (error) {
           logger.error({ err: error }, 'Failed to handle P2P chat entered');
         }
       },
-      // Issue #463: Handle bot added to group for welcome message
       'im.chat.member.added_v1': async (data: unknown) => {
         try {
-          await this.handleChatMemberAdded(data as FeishuChatMemberAddedEventData);
+          await this.welcomeHandler.handleChatMemberAdded(data as any);
         } catch (error) {
           logger.error({ err: error }, 'Failed to handle chat member added');
         }
@@ -228,12 +219,9 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
         );
         break;
       case 'file':
-        // TODO: Pass threadId when Issue #68 is implemented
         await sender.sendFile(message.chatId, message.filePath || '');
         break;
       case 'done':
-        // Task completion signal, no actual message to send
-        // This is used for REST sync mode and internal signaling
         logger.debug({ chatId: message.chatId }, 'Task completed (done signal)');
         break;
       default:
@@ -247,8 +235,6 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
 
   /**
    * Get the capabilities of Feishu channel.
-   * Feishu supports cards, threads, files, markdown, mentions, and updates.
-   * Issue #590 Phase 3: Added supportedMcpTools for dynamic prompt adaptation.
    */
   getCapabilities(): ChannelCapabilities {
     return {
@@ -269,7 +255,6 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
 
   /**
    * Get the TaskFlowOrchestrator for this channel.
-   * Used by deep-task skill MCP tools.
    */
   getTaskFlowOrchestrator(): TaskFlowOrchestrator | undefined {
     return this.taskFlowOrchestrator;
@@ -277,8 +262,6 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
 
   /**
    * Initialize TaskFlowOrchestrator with callbacks.
-   * Called by PrimaryNode after channel is created.
-   * Starts the file watcher to detect new Task.md files.
    */
   async initTaskFlowOrchestrator(callbacks: {
     sendMessage: (chatId: string, text: string) => Promise<void>;
@@ -290,7 +273,6 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
       callbacks,
       logger
     );
-    // Start the file watcher
     await this.taskFlowOrchestrator.start();
   }
 
@@ -304,7 +286,7 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
         client: this.client,
         logger,
       });
-      // Initialize filtered message forwarder (Issue #597)
+      // Initialize filtered message forwarder
       filteredMessageForwarder.setMessageSender({
         sendText: async (chatId: string, text: string) => {
           await this.messageSender!.sendText(chatId, text);
@@ -315,547 +297,85 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
   }
 
   /**
-   * Extract open_id from sender object.
+   * Create message handler context.
    */
-  private extractOpenId(sender?: { sender_type?: string; sender_id?: unknown }): string | undefined {
-    if (!sender?.sender_id) {
-      return undefined;
-    }
-    if (typeof sender.sender_id === 'object' && sender.sender_id !== null) {
-      const senderId = sender.sender_id as { open_id?: string };
-      return senderId.open_id;
-    }
-    if (typeof sender.sender_id === 'string') {
-      return sender.sender_id;
-    }
-    return undefined;
-  }
-
-  /**
-   * Add typing reaction to indicate processing started.
-   */
-  private async addTypingReaction(messageId: string): Promise<void> {
-    if (this.messageSender) {
-      await this.messageSender.addReaction(messageId, REACTIONS.TYPING);
-    }
-  }
-
-  /**
-   * Check if the chat is a group chat.
-   * Uses chat_type field from message event.
-   *
-   * @param chatType - Chat type from message event ('p2p', 'group', 'topic')
-   * @returns true if it's a group chat
-   */
-  private isGroupChat(chatType?: string): boolean {
-    return chatType === 'group' || chatType === 'topic';
-  }
-
-  /**
-   * Check if a chat ID is a group chat based on ID prefix.
-   * In Feishu, group chat IDs start with 'oc_' and private chat IDs start with 'ou_'.
-   *
-   * Issue #676: Used in handleChatMemberAdded where chat_type is not available.
-   *
-   * @param chatId - Chat ID to check
-   * @returns true if it's a group chat ID
-   */
-  private isGroupChatId(chatId: string): boolean {
-    return chatId.startsWith('oc_');
-  }
-
-  /**
-   * Fetch bot's open_id from Feishu API.
-   * This is used to correctly identify when the bot is mentioned.
-   *
-   * Issue #600: Correctly identify bot mentions in group chats
-   */
-  /**
-   * Fetch bot's info from Feishu API.
-   * This is used to correctly identify when the bot is mentioned.
-   *
-   * Issue #600: Correctly identify bot mentions in group chats
-   * Issue #681: Improve bot mention detection reliability
-   *
-   * @see https://open.feishu.cn/document/uAjLw4CM/ukTMukTMukTM/reference/bot-v3/bot_info/get
-   */
-  private async fetchBotInfo(): Promise<void> {
-    try {
-      const client = this.getClient();
-      // Use bot info API to get bot's open_id and app_id
-      const response = await client.request({
-        method: 'GET',
-        url: '/open-apis/bot/v3/info',
-      });
-
-      const bot = response.data?.bot;
-      if (bot?.open_id) {
-        this.botInfo = {
-          open_id: bot.open_id,
-          app_id: bot.app_id,
-        };
-        logger.info(
-          { botOpenId: bot.open_id, botAppId: bot.app_id },
-          'Bot info fetched for mention detection'
-        );
-      } else {
-        logger.warn('Failed to fetch bot info, mention detection may be less accurate');
-      }
-    } catch (error) {
-      logger.warn({ err: error }, 'Failed to fetch bot info, mention detection may be less accurate');
-    }
-  }
-
-  /**
-   * Check if the bot is mentioned in the message.
-   * When bot is mentioned, commands should be passed through to the agent.
-   *
-   * Issue #600: Correctly identify bot mentions in group chats
-   * Issue #681: Improve bot mention detection reliability
-   *
-   * Based on Feishu official documentation:
-   * - When bot is mentioned, mentions[].id.open_id may be bot's open_id OR app_id
-   * - We need to check both to ensure reliable detection
-   *
-   * @param mentions - Mentions array from Feishu message
-   * @returns true if bot is mentioned
-   */
-  private isBotMentioned(mentions?: FeishuMessageEvent['message']['mentions']): boolean {
-    if (!mentions || mentions.length === 0) {
-      return false;
-    }
-
-    // Log mentions structure for debugging
-    logger.debug(
-      {
-        mentions: JSON.stringify(mentions),
-        botInfo: this.botInfo,
+  private createMessageHandlerContext(): FeishuMessageHandlerContext {
+    return {
+      client: this.client!,
+      fileHandler: this.fileHandler,
+      mentionDetector: this.mentionDetector,
+      passiveModeManager: this.passiveModeManager,
+      appId: this.appId,
+      addTypingReaction: async (messageId: string) => {
+        if (this.messageSender) {
+          await this.messageSender.addReaction(messageId, REACTIONS.TYPING);
+        }
       },
-      'Checking bot mention'
-    );
-
-    // If we have bot info, check if any mention matches bot's open_id OR app_id
-    if (this.botInfo) {
-      return mentions.some((mention) => {
-        const mentionOpenId = mention.id?.open_id || '';
-        // Check against both bot's open_id and app_id
-        // Feishu may use either when the bot is mentioned
-        return (
-          mentionOpenId === this.botInfo!.open_id ||
-          mentionOpenId === this.botInfo!.app_id
-        );
-      });
-    }
-
-    // Fallback: Check for bot mention patterns
-    // Bot mentions typically have open_id starting with 'cli_' (app ID format)
-    // or have key containing 'bot'
-    return mentions.some((mention) => {
-      const openId = mention.id?.open_id || '';
-      const key = mention.key || '';
-      // Bot's open_id typically starts with 'cli_' (app/bot ID format)
-      // or the key contains 'bot' (e.g., '@_bot')
-      return openId.startsWith('cli_') || key.toLowerCase().includes('bot');
-    });
-  }
-
-  /**
-   * Check if passive mode is disabled for a specific chat.
-   * When passive mode is disabled, the bot responds to all messages in group chats.
-   *
-   * Issue #511: Group chat passive mode control
-   *
-   * @param chatId - Chat ID to check
-   * @returns true if passive mode is disabled (bot responds to all messages)
-   */
-  isPassiveModeDisabled(chatId: string): boolean {
-    return this.passiveModeDisabled.get(chatId) === true;
-  }
-
-  /**
-   * Set passive mode state for a specific chat.
-   *
-   * Issue #511: Group chat passive mode control
-   *
-   * @param chatId - Chat ID to configure
-   * @param disabled - true to disable passive mode (respond to all messages)
-   */
-  setPassiveModeDisabled(chatId: string, disabled: boolean): void {
-    if (disabled) {
-      this.passiveModeDisabled.set(chatId, true);
-      logger.info({ chatId }, 'Passive mode disabled for chat');
-    } else {
-      this.passiveModeDisabled.delete(chatId);
-      logger.info({ chatId }, 'Passive mode enabled for chat');
-    }
-  }
-
-  /**
-   * Get all chats with passive mode disabled.
-   *
-   * Issue #511: Group chat passive mode control
-   *
-   * @returns Array of chat IDs with passive mode disabled
-   */
-  getPassiveModeDisabledChats(): string[] {
-    return Array.from(this.passiveModeDisabled.keys());
-  }
-
-  /**
-   * Forward a filtered message to the debug chat.
-   * @see Issue #597
-   *
-   * @param reason - The reason the message was filtered
-   * @param messageId - The message ID
-   * @param chatId - The chat ID
-   * @param content - The message content
-   * @param userId - The user ID (optional)
-   * @param metadata - Additional metadata (optional)
-   */
-  private async forwardFilteredMessage(
-    reason: FilterReason,
-    messageId: string,
-    chatId: string,
-    content: string,
-    userId?: string,
-    metadata?: Record<string, unknown>
-  ): Promise<void> {
-    await filteredMessageForwarder.forward({
-      messageId,
-      chatId,
-      userId,
-      content,
-      reason,
-      metadata,
-      timestamp: Date.now(),
-    });
+      sendMessage: async (msg) => {
+        await this.sendMessage(msg as OutgoingMessage);
+      },
+      emitMessage: async (msg) => {
+        await this.emitMessage({
+          messageId: msg.messageId,
+          chatId: msg.chatId,
+          userId: msg.userId,
+          content: msg.content,
+          messageType: msg.messageType as any,
+          timestamp: msg.timestamp,
+          threadId: msg.threadId,
+          metadata: msg.metadata,
+          attachments: msg.attachments,
+        });
+      },
+      emitControl: async (ctrl) => {
+        return this.emitControl({
+          type: ctrl.type as any,
+          chatId: ctrl.chatId,
+          data: ctrl.data,
+        });
+      },
+      controlHandler: this.controlHandler,
+    };
   }
 
   /**
    * Handle incoming message event from WebSocket.
+   * This method delegates to FeishuMessageHandler for test compatibility.
+   * @internal Used by tests
    */
-  private async handleMessageReceive(data: FeishuEventData): Promise<void> {
-    if (!this.isRunning) {return;}
-
-    this.getClient();
-
-    const event = (data.event || data) as FeishuMessageEvent;
-    const { message, sender } = event;
-
-    if (!message) {return;}
-
-    const { message_id, chat_id, chat_type, content, message_type, create_time, mentions } = message;
-
-    // Bot replies to user message by setting parent_id = message_id
-    // Feishu automatically handles thread affiliation
-    const threadId = message_id;
-
-    if (!message_id || !chat_id || !content || !message_type) {
-      logger.warn('Missing required message fields');
-      return;
+  async handleMessageReceive(data: unknown): Promise<void> {
+    if (!this.client) {
+      this.getClient();
     }
-
-    // Deduplication
-    if (messageLogger.isMessageProcessed(message_id)) {
-      logger.debug({ messageId: message_id }, 'Skipped duplicate message');
-      await this.forwardFilteredMessage('duplicate', message_id, chat_id, content, this.extractOpenId(sender));
-      return;
-    }
-
-    // Ignore bot messages
-    if (sender?.sender_type === 'app') {
-      logger.debug('Skipped bot message');
-      await this.forwardFilteredMessage('bot', message_id, chat_id, content);
-      return;
-    }
-
-    // Check message age
-    if (create_time) {
-      const messageAge = Date.now() - create_time;
-      if (messageAge > this.MAX_MESSAGE_AGE) {
-        logger.debug({ messageId: message_id }, 'Skipped old message');
-        await this.forwardFilteredMessage('old', message_id, chat_id, content, this.extractOpenId(sender), { age: messageAge });
-        return;
-      }
-    }
-
-    // Handle file/image messages
-    if (message_type === 'image' || message_type === 'file' || message_type === 'media') {
-      logger.info(
-        { chatId: chat_id, messageType: message_type, messageId: message_id },
-        'Processing file/image message'
-      );
-      const result = await this.fileHandler.handleFileMessage(chat_id, message_type, content, message_id);
-      if (!result.success) {
-        logger.error(
-          { chatId: chat_id, messageType: message_type, messageId: message_id, error: result.error },
-          'File/image processing failed - detailed error'
-        );
-        await this.sendMessage({
-          chatId: chat_id,
-          type: 'text',
-          text: `❌ 处理${message_type === 'image' ? '图片' : '文件'}失败: ${result.error || '未知错误'}`,
-        });
-        return;
-      }
-
-      const attachments = attachmentManager.getAttachments(chat_id);
-      if (attachments.length > 0) {
-        const latestAttachment = attachments[attachments.length - 1];
-        const uploadPrompt = this.fileHandler.buildUploadPrompt(latestAttachment);
-
-        await messageLogger.logIncomingMessage(
-          message_id,
-          this.extractOpenId(sender) || 'unknown',
-          chat_id,
-          `[File uploaded: ${latestAttachment.fileName}]`,
-          message_type,
-          create_time
-        );
-
-        // Emit as incoming message
-        await this.emitMessage({
-          messageId: `${message_id}-file`,
-          chatId: chat_id,
-          userId: this.extractOpenId(sender),
-          content: uploadPrompt,
-          messageType: 'file',
-          timestamp: create_time,
-          threadId,
-          attachments: [{
-            fileName: latestAttachment.fileName || 'unknown',
-            filePath: latestAttachment.localPath || '',
-            mimeType: latestAttachment.mimeType,
-          }],
-        });
-      }
-      return;
-    }
-
-    // Handle text and post messages
-    if (message_type !== 'text' && message_type !== 'post') {
-      logger.debug({ messageType: message_type }, 'Skipped unsupported message type');
-      await this.forwardFilteredMessage('unsupported', message_id, chat_id, content, this.extractOpenId(sender), { messageType: message_type });
-      return;
-    }
-
-    // Parse content
-    let text = '';
-    try {
-      const parsed = JSON.parse(content);
-      if (message_type === 'text') {
-        text = parsed.text?.trim() || '';
-      } else if (message_type === 'post' && parsed.content && Array.isArray(parsed.content)) {
-        for (const row of parsed.content) {
-          if (Array.isArray(row)) {
-            for (const segment of row) {
-              if (segment?.tag === 'text' && segment.text) {
-                text += segment.text;
-              }
-            }
-          }
-        }
-        text = text.trim();
-      }
-    } catch {
-      logger.error('Failed to parse content');
-      return;
-    }
-
-    if (!text) {
-      logger.debug('Skipped empty text');
-      await this.forwardFilteredMessage('empty', message_id, chat_id, content, this.extractOpenId(sender));
-      return;
-    }
-
-    logger.info({ messageId: message_id, chatId: chat_id }, 'Message received');
-
-    // Log message
-    await messageLogger.logIncomingMessage(
-      message_id,
-      this.extractOpenId(sender) || 'unknown',
-      chat_id,
-      text,
-      message_type,
-      create_time
+    await this.feishuMessageHandler.handleMessageReceive(
+      data as any,
+      this.createMessageHandlerContext()
     );
-
-    // Check for control commands
-    // Control commands should ALWAYS be handled through the control channel, regardless of mentions
-    // This ensures /reset, /status, etc. work correctly even when bot is @mentioned
-    const botMentioned = this.isBotMentioned(mentions);
-
-    // Get control commands from CommandRegistry (Issue #463: removed hardcoded list)
-    const commandRegistry = getCommandRegistry();
-
-    // Issue #698: Strip leading mentions to detect commands in messages like "@bot /help"
-    // After stripping, we can check if the remaining text starts with '/'
-    const textWithoutMentions = stripLeadingMentions(text, mentions);
-
-    // Issue #460 & #511: Group chat passive mode
-    // In group chats, only respond when bot is mentioned (@bot)
-    // This allows scheduled tasks to broadcast without triggering unwanted responses
-    // Issue #511: Passive mode can be disabled per chat via /passive command
-    // Issue #650: Move passive mode check BEFORE command processing
-    // Issue #677: Allow /passive command to bypass passive mode check to avoid deadlock
-    // (when mention detection fails, users still need a way to disable passive mode)
-    const isPassiveCommand = textWithoutMentions.startsWith('/passive');
-    const passiveModeDisabled = this.isPassiveModeDisabled(chat_id);
-    if (this.isGroupChat(chat_type) && !botMentioned && !passiveModeDisabled && !isPassiveCommand) {
-      logger.debug(
-        { messageId: message_id, chatId: chat_id, chat_type },
-        'Skipped group chat message without @mention (passive mode)'
-      );
-      // Issue #597: Forward filtered message to debug chat
-      await this.forwardFilteredMessage('passive_mode', message_id, chat_id, text, this.extractOpenId(sender), { chat_type });
-      return;
-    }
-
-    if (textWithoutMentions.startsWith('/')) {
-      const [command, ...args] = textWithoutMentions.slice(1).split(/\s+/);
-      const cmd = command.toLowerCase();
-
-      // Handle control commands through the control channel
-      // Control commands are ALWAYS handled locally, regardless of @mentions
-      // Non-control commands with @mention show error instead of passing to agent (Issue #595)
-      const isControlCommand = commandRegistry.has(cmd);
-
-      if (isControlCommand || !botMentioned) {
-        if (this.controlHandler) {
-          const response = await this.emitControl({
-            type: cmd as any,
-            chatId: chat_id,
-            data: { args, rawText: textWithoutMentions, senderOpenId: this.extractOpenId(sender) },
-          });
-
-          // Only return if command was successfully handled
-          // Unknown commands (success: false) will fall through to normal message processing
-          if (response.success) {
-            if (response.message) {
-              await this.sendMessage({
-                chatId: chat_id,
-                type: 'text',
-                text: response.message,
-              });
-            }
-            return;
-          }
-          // Without @mention: unknown commands fall through to agent (original behavior)
-          // With @mention: show error instead of passing to agent (Issue #595)
-          if (botMentioned) {
-            await this.sendMessage({
-              chatId: chat_id,
-              type: 'text',
-              text: `❓ **未知命令**: /${cmd}\n\n使用 /help 查看可用命令列表。`,
-            });
-            return;
-          }
-        }
-
-        // Default command handling if no control handler registered
-        if (cmd === 'reset') {
-          await this.sendMessage({
-            chatId: chat_id,
-            type: 'text',
-            text: '✅ **对话已重置**\n\n新的会话已启动，之前的上下文已清除。',
-          });
-          return;
-        }
-
-        if (cmd === 'status') {
-          await this.sendMessage({
-            chatId: chat_id,
-            type: 'text',
-            text: `📊 **状态**\n\nChannel: ${this.name}\nStatus: ${this.status}`,
-          });
-          return;
-        }
-      } else {
-        // Unknown command with @mention: show error instead of passing to agent
-        // Issue #595: Control commands not parsed when bot is @mentioned in group chat
-        await this.sendMessage({
-          chatId: chat_id,
-          type: 'text',
-          text: `❓ **未知命令**: /${cmd}\n\n使用 /help 查看可用命令列表。`,
-        });
-        return;
-      }
-    }
-
-    // Log if bot is mentioned with a non-control command (for debugging)
-    if (botMentioned && textWithoutMentions.startsWith('/')) {
-      logger.debug({ messageId: message_id, chatId: chat_id, command: textWithoutMentions }, 'Bot mentioned with non-control command, passing to agent');
-    }
-
-    // Issue #514: Add typing reaction only for messages that will be processed
-    // This is placed after passive mode check to avoid reacting to skipped messages
-    await this.addTypingReaction(message_id);
-
-    // Issue #517: Get chat history for passive mode context
-    // When bot is mentioned in a group chat, include recent chat history as context
-    const isPassiveModeTrigger = this.isGroupChat(chat_type) && botMentioned;
-    let chatHistoryContext: string | undefined;
-
-    if (isPassiveModeTrigger) {
-      chatHistoryContext = await this.getChatHistoryContext(chat_id);
-      logger.debug(
-        { messageId: message_id, chatId: chat_id, historyLength: chatHistoryContext?.length },
-        'Including chat history context for passive mode trigger'
-      );
-    }
-
-    // Emit as incoming message
-    await this.emitMessage({
-      messageId: message_id,
-      chatId: chat_id,
-      userId: this.extractOpenId(sender),
-      content: text,
-      messageType: message_type as any,
-      timestamp: create_time,
-      threadId,
-      metadata: chatHistoryContext ? { chatHistoryContext } : undefined,
-    });
   }
 
   /**
-   * Get formatted chat history context for passive mode.
-   * Issue #517: Include recent chat history when bot is mentioned in group chats.
-   *
-   * @param chatId - Chat ID to get history for
-   * @returns Formatted chat history string, or undefined if no history
+   * Check if passive mode is disabled for a specific chat.
    */
-  private async getChatHistoryContext(chatId: string): Promise<string | undefined> {
-    try {
-      const rawHistory = await messageLogger.getChatHistory(chatId);
+  isPassiveModeDisabled(chatId: string): boolean {
+    return this.passiveModeManager.isDisabled(chatId);
+  }
 
-      if (!rawHistory || rawHistory.length === 0) {
-        return undefined;
-      }
+  /**
+   * Set passive mode state for a specific chat.
+   */
+  setPassiveModeDisabled(chatId: string, disabled: boolean): void {
+    this.passiveModeManager.setDisabled(chatId, disabled);
+  }
 
-      // Truncate if too long (keep the most recent content)
-      let history = rawHistory;
-      if (history.length > CHAT_HISTORY.MAX_CONTEXT_LENGTH) {
-        // Try to truncate at a reasonable point (e.g., at a message boundary)
-        const truncatePoint = history.lastIndexOf('## [', history.length - CHAT_HISTORY.MAX_CONTEXT_LENGTH);
-        if (truncatePoint > 0) {
-          history = `...(earlier messages truncated)...\n\n${history.slice(truncatePoint)}`;
-        } else {
-          // Fallback: just truncate from the end
-          history = history.slice(-CHAT_HISTORY.MAX_CONTEXT_LENGTH);
-          history = `...(earlier messages truncated)...\n\n${history.slice(history.indexOf('## ['))}`;
-        }
-      }
-
-      return history;
-    } catch (error) {
-      logger.error({ err: error, chatId }, 'Failed to get chat history context');
-      return undefined;
-    }
+  /**
+   * Get all chats with passive mode disabled.
+   */
+  getPassiveModeDisabledChats(): string[] {
+    return this.passiveModeManager.getDisabledChats();
   }
 
   /**
    * Handle card action event from WebSocket.
-   * Triggered when user clicks button, selects menu, etc. on an interactive card.
    */
   private async handleCardAction(data: FeishuCardActionEventData): Promise<void> {
     if (!this.isRunning) {
@@ -892,14 +412,10 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
 
     if (resolved) {
       logger.debug({ messageId: message_id }, 'Card action resolved pending interaction');
-      // Issue #657: Continue to emit message to agent instead of returning early
-      // This allows the agent to handle the interaction and decide what to do next
     }
 
-    // Issue #657: Always emit card action as a message to the agent
-    // This enables the agent to handle user interactions and take appropriate actions
+    // Always emit card action as a message to the agent
     try {
-      // Get button text for user-friendly message
       const buttonText = action.text || action.value;
       const messageContent = `User clicked '${buttonText}' button`;
 
@@ -925,16 +441,12 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
       logger.error({ err: error, messageId: message_id, chatId: chat_id }, 'Failed to emit card action message');
     }
 
-    // Return early if resolved - the wait_for_interaction tool already returned the result
     if (resolved) {
       return;
     }
 
     try {
-      // Try to handle via InteractionManager
       const handled = await this.interactionManager.handleAction(event, async (defaultEvent) => {
-        // Default handler: emit as interaction message
-        // Issue #525: Use button text to generate user-friendly prompt
         const buttonText = defaultEvent.action.text || defaultEvent.action.value;
         const messageContent = `The user clicked '${buttonText}' button`;
 
@@ -961,7 +473,6 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
     } catch (error) {
       logger.error({ err: error, messageId: message_id, chatId: chat_id }, 'Card action handler error');
 
-      // Notify user of the error
       await this.sendMessage({
         chatId: chat_id,
         type: 'text',
@@ -971,79 +482,7 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
   }
 
   /**
-   * Handle P2P chat entered event.
-   * Triggered when a user starts a private chat with the bot.
-   * Issue #463: Send welcome message on first private chat.
-   */
-  private async handleP2PChatEntered(data: FeishuP2PChatEnteredEventData): Promise<void> {
-    if (!this.isRunning || !this.welcomeService) {
-      return;
-    }
-
-    const { event } = data;
-    if (!event?.user?.open_id) {
-      logger.debug('P2P chat entered event missing user info');
-      return;
-    }
-
-    const userId = event.user.open_id;
-    logger.info({ userId }, 'P2P chat entered, sending welcome message');
-
-    await this.welcomeService.handleP2PChatEntered(userId);
-  }
-
-  /**
-   * Handle chat member added event.
-   * Triggered when members are added to a chat.
-   * Issue #463: Send welcome message when bot is added to a group.
-   * Issue #676: Send help message when users join a group that already has the bot.
-   */
-  private async handleChatMemberAdded(data: FeishuChatMemberAddedEventData): Promise<void> {
-    if (!this.isRunning || !this.welcomeService) {
-      return;
-    }
-
-    const { event } = data;
-    if (!event?.chat_id || !event?.members || event.members.length === 0) {
-      logger.debug('Chat member added event missing required fields');
-      return;
-    }
-
-    // Only send messages to group chats
-    if (!this.isGroupChatId(event.chat_id)) {
-      logger.debug({ chatId: event.chat_id }, 'Member added to non-group chat, skipping');
-      return;
-    }
-
-    // Check if the bot is among the added members
-    // Bot's member_id_type is "app_id" and member_id is the bot's app_id
-    const botMemberAdded = event.members.some(
-      (member) => member.member_id_type === 'app_id' && member.member_id === this.appId
-    );
-
-    // Get non-bot members (users who joined)
-    const userMembers = event.members.filter(
-      (member) => !(member.member_id_type === 'app_id' && member.member_id === this.appId)
-    );
-
-    if (botMemberAdded) {
-      // Bot was added to the group -> send welcome message
-      logger.info({ chatId: event.chat_id }, 'Bot added to group, sending welcome message');
-      await this.welcomeService.handleBotAddedToGroup(event.chat_id);
-    } else if (userMembers.length > 0) {
-      // Users joined a group that already has the bot -> send help message
-      logger.info(
-        { chatId: event.chat_id, userCount: userMembers.length },
-        'New users joined group, sending help message'
-      );
-      const userIds = userMembers.map((m) => m.member_id);
-      await this.welcomeService.handleUserJoinedGroup(event.chat_id, userIds);
-    }
-  }
-
-  /**
    * Get the InteractionManager for this channel.
-   * Used for registering custom interaction handlers.
    */
   getInteractionManager(): InteractionManager {
     return this.interactionManager;
@@ -1051,9 +490,8 @@ export class FeishuChannel extends BaseChannel<FeishuChannelConfig> {
 
   /**
    * Set the WelcomeService for this channel.
-   * Used for sending welcome messages on bot added to group or first private chat.
    */
   setWelcomeService(service: WelcomeService): void {
-    this.welcomeService = service;
+    this.welcomeHandler.setWelcomeService(service);
   }
 }
